requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.12
httpx==0.24.1
playwright==1.44.0
pandas==2.1.4
PyYAML==6.0
python-dotenv==1.0.0
PyPDF2==3.0.1
matplotlib==3.8.2
aiofiles==23.1.0
requests==2.31.0
